requirements.txt
pylint
pytest
streamlit
pandas
numpy
altair
more-itertools
streamlit-aggrid
plotly_express
requests
streamlit-lottie
datetime
